Criação de DTOs; Melhorias e alterações no controle global de exceção; Adição da validação;

+++ src/main/java/com/residencia/academia/controller/TurmaController.java
@@ -28,7 +28,7 @@ public class TurmaController {
 	public ResponseEntity<List<Turma>> findAllTurma() {
 		return new ResponseEntity<>(turmaService.findAllTurma(), HttpStatus.OK);
 	}
-
+	
 	@GetMapping("/{id}")
 	public ResponseEntity<Turma> findTurmaById(@PathVariable Integer id) {
 		Turma turma = turmaService.findTurmaById(id);
@@ -38,6 +38,15 @@ public class TurmaController {
 			return new ResponseEntity<>(turmaService.findTurmaById(id), HttpStatus.OK);
 	}
 
+	@GetMapping("/sem-verificacao/{id}")
+	public ResponseEntity<Turma> findTurmaByIdSemVerificacao(@PathVariable Integer id)
+			throws NoSuchElementFoundException{
+		
+		Turma turma = turmaService.findTurmaByIdSemVerificacao(id);
+		return new ResponseEntity<>(turma, HttpStatus.OK);
+	}
+	
+	
 	@PostMapping
 	public ResponseEntity<Turma> saveTurma(@RequestBody Turma turma) {
 		return new ResponseEntity<>(turmaService.saveTurma(turma), HttpStatus.CREATED);
@@ -47,11 +56,31 @@ public class TurmaController {
 	public ResponseEntity<Turma> updateTurma(@RequestBody Turma turma) {
 		return new ResponseEntity<>(turmaService.updateTurma(turma), HttpStatus.OK);
 	}
-
+	
 	@DeleteMapping("/{id}")
 	public ResponseEntity<String> deleteTurma(@PathVariable Integer id) {
+		Turma turma = turmaService.findTurmaById(id);
+		if(null == turma)
+			throw new NoSuchElementFoundException("Não foi possível excluir a Turma, "
+					+ "pois não foi "
+					+ "encontrada uma turma com o id " + id);
+
 		turmaService.deleteTurma(id);
 		return new ResponseEntity<>("", HttpStatus.OK);
 	}
-
+	
+	/*
+	@DeleteMapping("/{id}")
+	public ResponseEntity<String> deleteTurmaComConferencia(@PathVariable Integer id) {
+		Boolean verificacao = turmaService.deleteTurmaComConferencia(id);
+		if(verificacao)
+			return new ResponseEntity<>("", HttpStatus.OK);
+		else
+			throw new NoSuchElementFoundException("Não foi possível excluir a Turma, "
+					+ "pois não foi "
+					+ "encontrada uma turma com o id " + id);
+		
+	}
+	*/
+	
 }
